+++ docs/changelog.rst
@@ -13,6 +13,8 @@ Minor changes
 - Changed :meth:`spotify.Link.as_playlist()` to also support creating playlists
   from links with type :attr:`spotify.LinkType.STARRED`.
 
+- Changed from nose to py.test as test runner.
+
 Bug fixes
 ---------